requirements.txt
beautifulsoup4>=4.10.0
discord_py_interactions>=5.7.0
python-dotenv>=1.0.0
Requests>=2.31.0
kerykeion>=4.2.3
pycountry>=22.3.5
prettytable>=3.8.0
pandas>=2.1.0
timezonefinder>=6.2.0
geopy>=2.4.0